pyproject.toml
[project]
name = "tulun"
version = "1.0.0"
description = "Transparent and Adaptable Low-resource Machine Translation, through LLM post-editing"
authors = [
    {name = "Raphael Merx",email = "raphael.merx@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "daphne (>=4.1.2,<5.0.0)",
    "transformers (>=4.50.1,<5.0.0)",
    "google-cloud-translate (>=3.20.2,<4.0.0)",
    "spacy (>=3.8.4,<4.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "dspy (>=2.6.15,<3.0.0)",
    "tantivy (>=0.22.2,<0.23.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "litellm (>=1.62.0,<2.0.0)",
    "torch (>=2.6.0,<3.0.0) ; sys_platform == \"linux\"",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch_cpu", platform = "linux"}
en_core_web_sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz"}